requirements.txt
google-api-python-client>=2.120.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-generativeai>=0.8.0
pyyaml>=6.0
requests>=2.31.0
click>=8.1.7
python-dotenv>=1.0.0
